Added items

+++ db/migrate/20200708230314_add_total_to_trade.rb
@@ -1,5 +1,5 @@
 class AddTotalToTrade < ActiveRecord::Migration[6.0]
   def change
-    add_column :trades, :total, :integer, default: 0
+    add_column :trades, :total, :bigint, default: 0
   end
 end
